fix: move CLI provider selection into pkg/workers/cliprovider for pkg-file-count
