+++ pkg/workers/cliprovider/cli_provider_registry_test.go
@@ -1,4 +1,4 @@
-package workers
+package cliprovider
 
 import (
 	"errors"
