+++ pkg/workers/cliprovider/cli_provider_selection.go
@@ -1,4 +1,4 @@
-package workers
+package cliprovider
 
 import (
 	"fmt"
